# frontend/openapi.json
{
  "components": {
...
    "schemas": {
...
      "AlertList": {
        "properties": {
          "checked_at": {
            "description": "Server time the checks ran.",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Alert"
            },
            "type": "array"
          }
        },
        "required": [
          "items",
          "checked_at"
        ],
        "type": "object"
      },
...
      "ApiKeyCreated": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiKey"
          },
          {
            "properties": {
              "secret": {
                "description": "Shown this once - it cannot be read again.",
                "type": "string"
              }
            },
            "required": [
              "secret"
            ],
            "type": "object"
          }
        ]
      },
...
      "ApiKeyList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ApiKey"
            },
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "V1Error": {
        "properties": {
          "code": {
            "description": "Stable and machine-readable, e.g. query_not_found, param_invalid, table_not_allowed, rate_limited - every code is listed in API.md under Errors.",
            "type": "string"
          },
          "detail": {
            "description": "The database driver's own message, where there is one.",
            "type": "string"
          },
          "error": {
            "description": "Human-readable; wording may change between releases.",
            "type": "string"
          },
          "errors": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Per-parameter problems, when the parameter rules were invalid.",
            "type": "object"
          },
          "request_id": {
            "description": "Matches the X-Request-Id header and the server log.",
            "type": "string"
          },
          "retry_after": {
            "description": "Seconds to wait, when rate_limited.",
            "type": "integer"
          },
          "timeout": {
            "description": "The time limit that was exceeded, when query_timeout.",
            "type": "number"
          }
        },
        "required": [
          "error",
          "code",
          "request_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Run SQL against configured databases and get the results back over HTTP.",
    "title": "QueryAPIGate",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/v1/alerts": {
      "get": {
        "description": "**Experimental** - may change in any minor release, noted in the changelog.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertList"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "What needs attention now - expiring keys, failing connections, slow or failing queries, rate limits being hit - most severe first",
        "tags": [
          "Management API v1"
        ],
        "x-experimental": true
      }
    },
    "/api/v1/api-keys": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyList"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "List API keys (grants, status, expiry, last use, usage - never secrets)",
        "tags": [
          "Management API v1"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreated"
                }
              }
            },
            "description": "OK",
            "headers": {
              "ETag": {
                "description": "Send back as If-Match to make a change conditional.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Already exists"
          }
        },
        "summary": "Create an API key; the response carries its secret, shown this once",
        "tags": [
          "Management API v1"
        ]
      }
    },
...
    "/api/v1/audit": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Matches the time, actor or target.",
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLog"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Administrative changes, newest first",
        "tags": [
          "Management API v1"
        ]
      }
    },
    "/api/v1/cache/entries": {
      "delete": {
        "responses": {
          "204": {
            "description": "Cleared"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Evict every entry",
        "tags": [
          "Management API v1"
        ]
      },
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheEntryList"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Missing or invalid credentials"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not the admin key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1Error"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "What is in the response cache now, soonest to expire first",
        "tags": [
          "Management API v1"
        ]
      }
    },
...
  },
  "security": [
    {},
    {
      "ApiKey": []
    }
  ]
}